+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [3.1.0] - 2023-04-12
 ### Added
 - Abstract SyncTransforms function
+- IsOverlapingPoint function
 
 ## [3.0.0] - 2023-04-03
 ### Removed